requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Trading APIs
coinbase-advanced-py==1.3.0
py-clob-client==0.22.0
websockets==12.0

# Technical Analysis
pandas-ta==0.3.14b0
ta-lib==0.4.28

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
tensorflow==2.15.0
keras==2.15.0

# Data Storage
sqlalchemy==2.0.23
pyarrow==14.0.1

# Async & Networking
aiohttp==3.9.1
aiofiles==23.2.1

# Discord Integration
discord-webhook==1.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
colorama==0.4.6
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
